More organization and code cleanup. Made single post page more responsive to various screen sizes
