+++ components/pages/blog/post/ArticleHeader.jsx
@@ -1,5 +1,116 @@
 import Image from "next/image";
 import ShareSheet from "./ShareSheet";
+import styled from "styled-components";
+
+const ArticleHeaderWrap = styled.div`
+  .category.component {
+    margin-top: 0;
+    margin-bottom: 20px;
+  }
+
+  .category-eyebrow__category,
+  .category-eyebrow__date {
+    display: block;
+  }
+
+  .category-eyebrow {
+    font-size: 12px;
+    line-height: 1.33337;
+    font-weight: 700;
+    letter-spacing: -0.01em;
+    font-family: -apple-system, BlinkMacSystemFont, 'SF Pro Text', 'SF Pro Icons', 'Helvetica Neue', 'Helvetica',
+      'Arial', sans-serif;
+  }
+
+  .category-eyebrow__date {
+    margin-top: 4px;
+  }
+
+  .category-eyebrow__date {
+    font-size: 14px;
+    line-height: 1.28577;
+    font-weight: 600;
+    letter-spacing: -0.016em;
+    font-family: -apple-system, BlinkMacSystemFont, 'SF Pro Text', 'SF Pro Icons', 'Helvetica Neue', 'Helvetica',
+      'Arial', sans-serif;
+  }
+
+  .pagetitle.component {
+    margin-top: 0;
+    margin-bottom: 0;
+  }
+
+  .pagetitle .hero-headline {
+    font-size: 48px;
+    line-height: 1.08349;
+    font-weight: 700;
+    letter-spacing: -0.003em;
+    font-family: -apple-system, BlinkMacSystemFont, 'SF Pro Display', 'SF Pro Icons', 'Helvetica Neue', 'Helvetica',
+      'Arial', sans-serif;
+
+    @media only screen and (max-width: 1068px) {
+      font-size: 40px;
+      line-height: 1.1;
+      letter-spacing: 0em;
+    }
+
+    @media only screen and (max-width: 734px) {
+      font-size: 32px;
+      line-height: 1.125;
+      letter-spacing: 0.004em;;
+    }
+  }
+
+  .article-subhead.component {
+    font-size: 24px;
+    line-height: 1.16667;
+    font-weight: 500;
+    letter-spacing: 0.009em;
+    font-family: -apple-system, BlinkMacSystemFont, 'SF Pro Display', 'SF Pro Icons', 'Helvetica Neue', 'Helvetica',
+      'Arial', sans-serif;
+    margin-top: 20px;
+    margin-bottom: 0;
+
+    @media only screen and (max-width: 1068px) {
+      font-size: 21px;
+      line-height: 1.19048;
+      letter-spacing: .011em;
+    }
+  }
+
+  .author.component {
+    margin-top: 20px;
+    margin-bottom: 0;
+  }
+
+  .author-content {
+    display: flex;
+    align-items: center;
+    gap: 5px;
+  }
+
+  .author-image {
+    width: 24px;
+    height: 24px;
+    border-radius: 50%;
+  }
+
+  .author-text {
+    font-size: 14px;
+    line-height: 1.28577;
+    font-weight: 600;
+    letter-spacing: -0.016em;
+    font-family: -apple-system, BlinkMacSystemFont, 'SF Pro Text', 'SF Pro Icons', 'Helvetica Neue', 'Helvetica',
+      'Arial', sans-serif;
+  }
+
+  @media only screen and (max-width: 734px) {
+    margin-left: auto;
+    margin-right: auto;
+    width: 87.5%;
+  }
+`;
+
 
 function formatDate(fullDate) {
   const date = new Date(fullDate);
@@ -11,9 +122,8 @@ function formatDate(fullDate) {
 }
 
 const ArticleHeader = ({ frontmatter, author }) => {
-  console.log("ArticleHeader:", author)
   return (
-    <div className="article-header">
+    <ArticleHeaderWrap className="article-header">
       <div className="category component">
         <div className="component-content">
           <div className="category-eyebrow">
@@ -53,7 +163,7 @@ const ArticleHeader = ({ frontmatter, author }) => {
         </div>
       )}
       <ShareSheet />
-    </div>
+    </ArticleHeaderWrap>
   );
 };
